feat: add stateless bottom navigation bar
