+++ lib/business_logic/view_models/bottom_navigation_bar_viewmodel.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/foundation.dart';
+
+class BottomNavigationBarViewModel extends ChangeNotifier {
+  int _selectedIndex = 0;
+
+  //Get current selected screen index
+  int get currentSelectedIndex => _selectedIndex;
+
+  //Update selected screen index
+  void updateSelectedIndex(int newIndex) {
+    _selectedIndex = newIndex;
+    notifyListeners();
+  }
+}
