+++ lib/business_logic/view_models/viewmodels.dart
@@ -1,3 +1,4 @@
 export './movie_viewmodel.dart';
 export './home_viewmodel.dart';
 export './movie_detail_viewmodel.dart';
+export './bottom_navigation_bar_viewmodel.dart';
