+++ lib/ui/widgets/widgets.dart
@@ -2,3 +2,4 @@ export './theme_switch.dart';
 export './movie_card.dart';
 export './user_app_profile.dart';
 export './custom_app_bar.dart';
+export './bottom_navigation_bar.dart';
