+++ pubspec.yaml
@@ -14,6 +14,7 @@ dependencies:
   flutter_dotenv: ^5.0.2
   get_it: ^7.2.0
   google_fonts: ^3.0.1
+  provider: ^6.0.3
 
 dev_dependencies:
   flutter_test:
